requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
apscheduler==3.9.1
requests==2.28
factory-boy==3.2.1